printing decimals

+++ main.h
@@ -2,6 +2,7 @@
 #define MAIN_H
 #include <stddef.h>
 #include <stdarg.h>
+/*#define l ((INT_MAX) + (1024))*/
 
 /**
  * struct myfunc - takes specifier and points to a function specified
@@ -21,5 +22,5 @@ int print_s(va_list list);
 int (*compare_func(const char *a))(va_list);
 int percentage_sign(va_list list __attribute__((unused)));
 int print_d(va_list list);
-void print_decimal(int n);
+void print_decimal(long int n);
 #endif
